+++ Z/IO/FileSystem/Threaded.hs
@@ -42,6 +42,7 @@ module Z.IO.FileSystem.Threaded
   , utime, futime, lutime
   , link, symlink
   , readlink, realpath
+  , chown, fchown, lchown
   -- * opening constant
   -- ** AccessMode
   , AccessMode
@@ -127,17 +128,17 @@ import           Z.IO.UV.Manager
 -- Implict offset interface is provided by 'Input' \/ 'Output' instances.
 -- Explict offset interface is provided by 'readFileT' \/ 'writeFileT'.
 --
-data FileT =  FileT  {-# UNPACK #-} !UVFD      -- ^ the file
+data FileT =  FileT  {-# UNPACK #-} !FD      -- ^ the file
                      {-# UNPACK #-} !(IORef Bool)  -- ^ closed flag
 
 -- | Return FileT fd.
-getFileTFD :: FileT -> IO UVFD
+getFileTFD :: FileT -> IO FD
 getFileTFD (FileT fd closedRef) = do
     closed <- readIORef closedRef
     if closed then throwECLOSED else return fd
 
 -- | If fd is -1 (closed), throw 'ResourceVanished' ECLOSED.
-checkFileTClosed :: HasCallStack => FileT -> (UVFD -> IO a) -> IO a
+checkFileTClosed :: HasCallStack => FileT -> (FD -> IO a) -> IO a
 checkFileTClosed (FileT fd closedRef) f = do
     closed <- readIORef closedRef
     if closed then throwECLOSED else f fd
@@ -540,3 +541,23 @@ realpath path = do
                     (\ _ -> hs_uv_fs_readlink_extra_cleanup p'))
         (hs_uv_fs_readlink_cleanup . fst)
         (fromCString . fst)
+
+-- | Equivalent to <http://linux.die.net/man/2/chown chown(2)>.
+chown :: HasCallStack => CBytes -> UID -> GID -> IO ()
+chown path uid gid = do
+    uvm <- getUVManager
+    withCBytesUnsafe path $ \ p ->
+        withUVRequest_ uvm (hs_uv_fs_chown_threaded p uid gid)
+
+-- | Equivalent to <http://linux.die.net/man/2/fchown fchown(2)>.
+fchown :: HasCallStack => FD -> UID -> GID -> IO ()
+fchown fd uid gid = do
+    uvm <- getUVManager
+    withUVRequest_ uvm (hs_uv_fs_fchown_threaded fd uid gid)
+
+-- | Equivalent to <http://linux.die.net/man/2/lchown lchown(2)>.
+lchown :: HasCallStack => CBytes -> UID -> GID -> IO ()
+lchown path uid gid = do
+    uvm <- getUVManager
+    withCBytesUnsafe path $ \ p ->
+        withUVRequest_ uvm (hs_uv_fs_lchown_threaded p uid gid)